creating ui state
